fix: replace temporary links with links from data
