+++ src/components/authorPage/AuthorPage.jsx
@@ -4,25 +4,16 @@ import TimeLine from '../timeLine/TimeLine';
 import AuthorCutaway from '../authorCutaway/AuthorCutaway';
 import AuthorWorks from '../authorWorks/AuthorWorks';
 import './authorPage.css';
-import click from './overlay';
 import PhotoGallery from '../photoGallery/PhotoGallery';
-
-const photos = [
-  'http://catalog.library.mogilev.by/kray/Culture/photo/r5p15_1.jpg',
-  'https://i.livelib.ru/boocover/1002164588/200/6ffd/Ya%D1%9Egen_Guchok__Algerd_Abuhovich._Zacharavany_belaruskim_slovam.jpg',
-];
+import VideoOverlay from '../videoOverlay/VideoOverlay';
 
 const AuthorPage = ({ info }) => (
   <div className="author-page">
     <AuthorCutaway firstName={info.firstName} lastName={info.lastName} avatar={info.titleImage} />
     <AuthorWorks works={info.works} />
     <TimeLine className="timeline" info={info.timeline} />
-    <div className="photo-list"><PhotoGallery photos={photos} /></div>
-    <div className="video-list">
-      <img className="videosnap" alt="snap" src="#" onClick={click}/>
-      <div className="video-overlay"></div>
-      <div className="video-video"></div>
-    </div>
+    <PhotoGallery photos={info.images} />
+    <VideoOverlay video={info.videos} videosnap={info.videosnap} />
     <div className="map-component">map-component</div>
   </div>
 );
