+++ src/components/videoOverlay/videoOverlay.css
@@ -1,0 +1,23 @@
+.video-list {
+  position:relative;
+  text-align: center;
+}
+
+.video-overlay {
+  position:fixed;
+  top:0;
+  left:0;
+  z-index:1;
+  height:100%;
+  width:100%;
+  background:#ffffff;
+  opacity:0.6;
+  display:none;    
+}
+.video-video {
+  display:none;
+  position: absolute;
+  top:-315px;
+  left:-150px;
+  z-index:2;
+}
